prints the size of various types[fix]

+++ 0x00-hello_world/6-size.c
@@ -2,13 +2,13 @@
 /**
  * main - Entry point
  *
- * Return: Always 0 (Success)
+ * Return: If is 0 Success, else is not succes
  */
 int main(void)
 {
 printf("Size of a char: %d byte(s)\n", sizeof(char));
 printf("Size of an int: %d byte(s)\n", sizeof(int));
-printf("Size of a long int: %d byte(s)\n", sizeof(long long));
+printf("Size of a long int: %d byte(s)\n", sizeof(long int));
 printf("Size of a long long int: %d byte(s)\n", sizeof(long long int));
 printf("Size of a float: %d byte(s)\n", sizeof(float));
 return (0);
